% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultibarHorizontal.R
\name{MultibarHorizontal}
\alias{MultibarHorizontal}
\title{Multi-Bar Horizontal Plot in R}
\usage{
MultibarHorizontal(df, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{df}{json data, which looks like the one in examples/reg_results.txt (see GitHub). Please create a file and read from it.}
}
\description{
Visualizes horizontal bars, grouped and stacked.
}
\references{
Kolpashnikova, Kamila. (2022). Multi-Bar Horizontal Plot in R. Toronto,ON: York University.
}
